ddl: migrate test-infra to testify for index modify tests
