+++ ddl/db_test.go
@@ -19,7 +19,6 @@ import (
 	"context"
 	"fmt"
 	"math"
-	"math/rand"
 	"sort"
 	"strings"
 	"sync"
@@ -953,68 +952,6 @@ func (s *testDBSuite1) TestCancelAddTableAndDropTablePartition(c *C) {
 		}
 	}
 	s.dom.DDL().SetHook(originalHook)
-}
-
-func (s *testDBSuite1) TestDropPrimaryKey(c *C) {
-	idxName := "primary"
-	createSQL := "create table test_drop_index (c1 int, c2 int, c3 int, unique key(c1), primary key(c3) nonclustered)"
-	dropIdxSQL := "alter table test_drop_index drop primary key;"
-	testDropIndex(c, s.store, s.lease, createSQL, dropIdxSQL, idxName)
-}
-
-func (s *testDBSuite2) TestDropIndex(c *C) {
-	idxName := "c3_index"
-	createSQL := "create table test_drop_index (c1 int, c2 int, c3 int, unique key(c1), key c3_index(c3))"
-	dropIdxSQL := "alter table test_drop_index drop index c3_index;"
-	testDropIndex(c, s.store, s.lease, createSQL, dropIdxSQL, idxName)
-}
-
-func testDropIndex(c *C, store kv.Storage, lease time.Duration, createSQL, dropIdxSQL, idxName string) {
-	tk := testkit.NewTestKit(c, store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists test_drop_index")
-	tk.MustExec(createSQL)
-	done := make(chan error, 1)
-	tk.MustExec("delete from test_drop_index")
-
-	num := 100
-	// add some rows
-	for i := 0; i < num; i++ {
-		tk.MustExec("insert into test_drop_index values (?, ?, ?)", i, i, i)
-	}
-	ctx := tk.Se.(sessionctx.Context)
-	indexID := testGetIndexID(c, ctx, "test_db", "test_drop_index", idxName)
-	jobIDExt, reset := setupJobIDExtCallback(ctx)
-	defer reset()
-	testddlutil.SessionExecInGoroutine(store, "test_db", dropIdxSQL, done)
-
-	ticker := time.NewTicker(lease / 2)
-	defer ticker.Stop()
-LOOP:
-	for {
-		select {
-		case err := <-done:
-			if err == nil {
-				break LOOP
-			}
-			c.Assert(err, IsNil, Commentf("err:%v", errors.ErrorStack(err)))
-		case <-ticker.C:
-			step := 5
-			// delete some rows, and add some data
-			for i := num; i < num+step; i++ {
-				n := rand.Intn(num)
-				tk.MustExec("update test_drop_index set c2 = 1 where c1 = ?", n)
-				tk.MustExec("insert into test_drop_index values (?, ?, ?)", i, i, i)
-			}
-			num += step
-		}
-	}
-
-	rows := tk.MustQuery("explain select c1 from test_drop_index where c3 >= 0")
-	c.Assert(strings.Contains(fmt.Sprintf("%v", rows), idxName), IsFalse)
-
-	checkDelRangeAdded(tk, jobIDExt.jobID, indexID)
-	tk.MustExec("drop table test_drop_index")
 }
 
 func testGetIndexID(c *C, ctx sessionctx.Context, dbName, tblName, idxName string) int64 {
@@ -4899,227 +4836,6 @@ func (s *testDBSuite8) TestDdlMaxLimitOfIdentifier(c *C) {
 	// alter table
 	tk.MustExec(fmt.Sprintf("alter table %s change f2 %s int", longTblName, strings.Repeat("二", mysql.MaxColumnNameLength-1)))
 
-}
-
-func testDropIndexes(c *C, store kv.Storage, lease time.Duration, createSQL, dropIdxSQL string, idxNames []string) {
-	tk := testkit.NewTestKit(c, store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists test_drop_indexes")
-	tk.MustExec(createSQL)
-	done := make(chan error, 1)
-
-	num := 100
-	// add some rows
-	for i := 0; i < num; i++ {
-		tk.MustExec("insert into test_drop_indexes values (?, ?, ?)", i, i, i)
-	}
-	ctx := tk.Se.(sessionctx.Context)
-	idxIDs := make([]int64, 0, 3)
-	for _, idxName := range idxNames {
-		idxIDs = append(idxIDs, testGetIndexID(c, ctx, "test_db", "test_drop_indexes", idxName))
-	}
-	jobIDExt, resetHook := setupJobIDExtCallback(ctx)
-	defer resetHook()
-	testddlutil.SessionExecInGoroutine(store, "test_db", dropIdxSQL, done)
-
-	ticker := time.NewTicker(lease / 2)
-	defer ticker.Stop()
-LOOP:
-	for {
-		select {
-		case err := <-done:
-			if err == nil {
-				break LOOP
-			}
-			c.Assert(err, IsNil, Commentf("err:%v", errors.ErrorStack(err)))
-		case <-ticker.C:
-			step := 5
-			// delete some rows, and add some data
-			for i := num; i < num+step; i++ {
-				n := rand.Intn(num)
-				tk.MustExec("update test_drop_indexes set c2 = 1 where c1 = ?", n)
-				tk.MustExec("insert into test_drop_indexes values (?, ?, ?)", i, i, i)
-			}
-			num += step
-		}
-	}
-	for _, idxID := range idxIDs {
-		checkDelRangeAdded(tk, jobIDExt.jobID, idxID)
-	}
-}
-
-func testCancelDropIndexes(c *C, store kv.Storage, d ddl.DDL) {
-	indexesName := []string{"idx_c1", "idx_c2"}
-	addIdxesSQL := "alter table t add index idx_c1 (c1);alter table t add index idx_c2 (c2);"
-	dropIdxesSQL := "alter table t drop index idx_c1;alter table t drop index idx_c2;"
-
-	tk := testkit.NewTestKit(c, store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(c1 int, c2 int)")
-	defer tk.MustExec("drop table t;")
-	for i := 0; i < 5; i++ {
-		tk.MustExec("insert into t values (?, ?)", i, i)
-	}
-	testCases := []struct {
-		needAddIndex   bool
-		jobState       model.JobState
-		JobSchemaState model.SchemaState
-		cancelSucc     bool
-	}{
-		// model.JobStateNone means the jobs is canceled before the first run.
-		// if we cancel successfully, we need to set needAddIndex to false in the next test case. Otherwise, set needAddIndex to true.
-		{true, model.JobStateNone, model.StateNone, true},
-		{false, model.JobStateRunning, model.StateWriteOnly, false},
-		{true, model.JobStateRunning, model.StateDeleteOnly, false},
-		{true, model.JobStateRunning, model.StateDeleteReorganization, false},
-	}
-	var checkErr error
-	hook := &ddl.TestDDLCallback{}
-	var jobID int64
-	testCase := &testCases[0]
-	hook.OnJobRunBeforeExported = func(job *model.Job) {
-		if (job.Type == model.ActionDropIndex || job.Type == model.ActionDropPrimaryKey) &&
-			job.State == testCase.jobState && job.SchemaState == testCase.JobSchemaState {
-			jobID = job.ID
-			jobIDs := []int64{job.ID}
-			hookCtx := mock.NewContext()
-			hookCtx.Store = store
-			err := hookCtx.NewTxn(context.TODO())
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			txn, err := hookCtx.Txn(true)
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-
-			errs, err := admin.CancelJobs(txn, jobIDs)
-			if err != nil {
-				checkErr = errors.Trace(err)
-				return
-			}
-			if errs[0] != nil {
-				checkErr = errors.Trace(errs[0])
-				return
-			}
-			checkErr = txn.Commit(context.Background())
-		}
-	}
-	originalHook := d.GetHook()
-	d.SetHook(hook)
-	ctx := tk.Se.(sessionctx.Context)
-	for i := range testCases {
-		testCase = &testCases[i]
-		if testCase.needAddIndex {
-			tk.MustExec(addIdxesSQL)
-		}
-		rs, err := tk.Exec(dropIdxesSQL)
-		if rs != nil {
-			rs.Close()
-		}
-		t := testGetTableByName(c, ctx, "test_db", "t")
-
-		var indexInfos []*model.IndexInfo
-		for _, idxName := range indexesName {
-			indexInfo := t.Meta().FindIndexByName(idxName)
-			if indexInfo != nil {
-				indexInfos = append(indexInfos, indexInfo)
-			}
-		}
-
-		if testCase.cancelSucc {
-			c.Assert(checkErr, IsNil)
-			c.Assert(err, NotNil)
-			c.Assert(err.Error(), Equals, "[ddl:8214]Cancelled DDL job")
-			c.Assert(indexInfos, NotNil)
-			c.Assert(indexInfos[0].State, Equals, model.StatePublic)
-		} else {
-			err1 := admin.ErrCannotCancelDDLJob.GenWithStackByArgs(jobID)
-			c.Assert(err, IsNil)
-			c.Assert(checkErr, NotNil)
-			c.Assert(checkErr.Error(), Equals, err1.Error())
-			c.Assert(indexInfos, IsNil)
-		}
-	}
-	d.SetHook(originalHook)
-	tk.MustExec(addIdxesSQL)
-	tk.MustExec(dropIdxesSQL)
-}
-
-func testDropIndexesIfExists(c *C, store kv.Storage) {
-	tk := testkit.NewTestKitWithInit(c, store)
-	tk.MustExec("use test_db;")
-	tk.MustExec("drop table if exists test_drop_indexes_if_exists;")
-	tk.MustExec("create table test_drop_indexes_if_exists (id int, c1 int, c2 int, primary key(id), key i1(c1), key i2(c2));")
-
-	// Drop different indexes.
-	tk.MustGetErrMsg(
-		"alter table test_drop_indexes_if_exists drop index i1, drop index i3;",
-		"[ddl:1091]index i3 doesn't exist",
-	)
-	if _, err := tk.Exec("alter table test_drop_indexes_if_exists drop index i1, drop index if exists i3;"); true {
-		c.Assert(err, IsNil)
-	}
-	tk.MustQuery("show warnings;").Check(
-		testutil.RowsWithSep("|", "Warning|1091|index i3 doesn't exist"),
-	)
-
-	// Verify the impact of deletion order when dropping duplicate indexes.
-	tk.MustGetErrMsg(
-		"alter table test_drop_indexes_if_exists drop index i2, drop index i2;",
-		"[ddl:1091]index i2 doesn't exist",
-	)
-	tk.MustGetErrMsg(
-		"alter table test_drop_indexes_if_exists drop index if exists i2, drop index i2;",
-		"[ddl:1091]index i2 doesn't exist",
-	)
-	if _, err := tk.Exec("alter table test_drop_indexes_if_exists drop index i2, drop index if exists i2;"); true {
-		c.Assert(err, IsNil)
-	}
-	tk.MustQuery("show warnings;").Check(
-		testutil.RowsWithSep("|", "Warning|1091|index i2 doesn't exist"),
-	)
-}
-
-func testDropIndexesFromPartitionedTable(c *C, store kv.Storage) {
-	tk := testkit.NewTestKitWithInit(c, store)
-	tk.MustExec("use test_db;")
-	tk.MustExec("drop table if exists test_drop_indexes_from_partitioned_table;")
-	tk.MustExec(`
-		create table test_drop_indexes_from_partitioned_table (id int, c1 int, c2 int, primary key(id), key i1(c1), key i2(c2))
-		partition by range(id) (partition p0 values less than (6), partition p1 values less than maxvalue);
-	`)
-	for i := 0; i < 20; i++ {
-		tk.MustExec("insert into test_drop_indexes_from_partitioned_table values (?, ?, ?)", i, i, i)
-	}
-	if _, err := tk.Exec("alter table test_drop_indexes_from_partitioned_table drop index i1, drop index if exists i2;"); true {
-		c.Assert(err, IsNil)
-	}
-}
-
-func (s *testDBSuite5) TestDropIndexes(c *C) {
-	// drop multiple indexes
-	createSQL := "create table test_drop_indexes (id int, c1 int, c2 int, primary key(id), key i1(c1), key i2(c2));"
-	dropIdxSQL := "alter table test_drop_indexes drop index i1, drop index i2;"
-	idxNames := []string{"i1", "i2"}
-	testDropIndexes(c, s.store, s.lease, createSQL, dropIdxSQL, idxNames)
-
-	createSQL = "create table test_drop_indexes (id int, c1 int, c2 int, primary key(id) nonclustered, unique key i1(c1), key i2(c2));"
-	dropIdxSQL = "alter table test_drop_indexes drop primary key, drop index i1;"
-	idxNames = []string{"primary", "i1"}
-	testDropIndexes(c, s.store, s.lease, createSQL, dropIdxSQL, idxNames)
-
-	createSQL = "create table test_drop_indexes (uuid varchar(32), c1 int, c2 int, primary key(uuid), unique key i1(c1), key i2(c2));"
-	dropIdxSQL = "alter table test_drop_indexes drop primary key, drop index i1, drop index i2;"
-	idxNames = []string{"primary", "i1", "i2"}
-	testDropIndexes(c, s.store, s.lease, createSQL, dropIdxSQL, idxNames)
-
-	testDropIndexesIfExists(c, s.store)
-	testDropIndexesFromPartitionedTable(c, s.store)
-	testCancelDropIndexes(c, s.store, s.dom.DDL())
 }
 
 // Close issue #24580.
